add to output now in session scope
